Inventory_Management/requirements.txt
asgiref
certifi
charset-normalizer
Django
django-leaflet
djangorestframework
idna
pillow
psycopg2-binary
requests
sqlparse
typing_extensions
urllib3
